page 조회 api의 Pageable 파라미터 개선

+++ src/main/java/com/didacto/repository/lecture/LectureCustomRepositoryImpl.java
@@ -32,7 +32,7 @@ public class LectureCustomRepositoryImpl implements LectureCustomRepository {
             query.orderBy(new OrderSpecifier<>(order.isAscending() ? Order.ASC : Order.DESC, pathBuilder.get(order.getProperty())));
         }
 
-        long offset = (pageable.getPageNumber() - 1) * pageable.getPageSize(); // 페이지네이션 Offset 계산
+        int offset = (pageable.getPageNumber() - 1) * pageable.getPageSize(); // 페이지네이션 Offset 계산
 
         return query
                 .offset(offset)
